move services and routes into controllers
